requirements.txt
flask==2.2.3
gunicorn==20.1.0
python-dotenv==1.0.0